feat: command now uses enum not hardcoded values

+++ src-client/connection.cpp
@@ -148,7 +148,7 @@ int start_of_connection(Connection* conn) {
     }
     
     // send device name length and name
-    uint32_t name_len = htonl(static_cast<uint32_t>(device_name.size()));
+    uint64_t name_len = htonl(static_cast<uint64_t>(device_name.size()));
     if (safe_SSL_write(conn, &name_len, sizeof(name_len)) < 0) {
         std::cerr << "Failed to send device name length\n";
         return -1;
@@ -162,15 +162,15 @@ int start_of_connection(Connection* conn) {
 
 int end_of_connection(Connection* conn) {
     if (!conn) return -1;
-    uint32_t num_events_net;
+    uint64_t num_events_net;
     if (safe_SSL_read(conn, &num_events_net, sizeof(num_events_net)) <= 0) {
         std::cerr << "Failed to read number of pending events from server\n";
         close_connection(conn);
         return -1;
     }
-    uint32_t num_events = ntohl(num_events_net);
+    uint64_t num_events = ntohl(num_events_net);
     pending_events.store(num_events);
-    for (uint32_t i = 0; i < num_events && i < num_threads; ++i) {
+    for (uint64_t i = 0; i < num_events && i < num_threads; ++i) {
         events_cv.notify_one();
     }
     return close_connection(conn);
