+++ src-server/server.cpp
@@ -148,7 +148,7 @@ int main(int argc, char *argv[]) {
         
         std::thread([conn]() {
             // read client name length and name
-            uint32_t client_name_len = 0;
+            uint64_t client_name_len = 0;
             if (safe_SSL_read(conn, &client_name_len, sizeof(client_name_len)) <= 0) {
                 std::cerr << "Failed to read client name length\n";
                 close_connection(conn);
@@ -198,8 +198,8 @@ int main(int argc, char *argv[]) {
 
 void end_of_connection(Connection* conn, Event& event) {
     // send number of events that the client need to process
-    uint32_t num_events = get_pending_event_count(event.client_id);
-    uint32_t num_events_net = htonl(num_events);
+    uint64_t num_events = get_pending_event_count(event.client_id);
+    uint64_t num_events_net = htonl(num_events);
     if (safe_SSL_write(conn, &num_events_net, sizeof(num_events_net)) < 0) {
         std::cerr << "Failed to send number of pending events to client: " << event.client_id << "\n";
     }
